test: add upgrade tx + set SL chain ID service tx integration test
Move the service-after-upgrade test to the interop crate where the
system context contract bytecodes are available, and verify that the
settlement layer chain ID is actually updated.
